requirements.txt
asgiref==3.11.1
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
dj-database-url==3.1.0
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
et_xmlfile==2.0.0
inflection==0.5.1
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
pycparser==3.0
PyJWT==2.11.0
PyMySQL==1.1.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
reportlab==4.4.9
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
